Pulled user into its own directories. Added morgan logging. Added some comments.
