+++ package.json
@@ -21,6 +21,7 @@
     "cookie-parser": "^1.3.5",
     "express": "^4.12.4",
     "mongoose": "^4.0.5",
+    "morgan": "^1.6.0",
     "passport": "^0.2.2",
     "passport-local": "^1.0.0",
     "passport-local-mongoose": "^1.0.0"
